Fixing interface for TroopStub after merge

+++ Game/Data/Troop/ITroopStub.cs
@@ -32,7 +32,7 @@ namespace Game.Data.Troop
 
         ushort TotalCount { get; }
 
-        int TotalHp { get; }
+        decimal TotalHp { get; }
 
         byte Speed { get; }
